+++ tests/test_httpx_sync.py
@@ -610,9 +610,9 @@ def test_requests_json_body(httpx_mock: HTTPXMock):
 
 
 def test_callback_raising_exception(httpx_mock: HTTPXMock):
-    def raise_timeout(request, ext):
+    def raise_timeout(request, extensions):
         raise httpx.ReadTimeout(
-            f"Unable to read within {ext['timeout']['read']}", request=request
+            f"Unable to read within {extensions['timeout']['read']}", request=request
         )
 
     httpx_mock.add_callback(raise_timeout, url="http://test_url")