+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.12.0] - 2021-04-27
+### Changed
+- Requires [`httpx`](https://www.python-httpx.org)==0.18.\*
+- `ext` callback parameter was renamed into `extensions`.
+
 ## [0.11.0] - 2021-03-01
 ### Changed
 - Requires [`httpx`](https://www.python-httpx.org)==0.17.\*
@@ -127,7 +132,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - First release, should be considered as unstable for now as design might change.
 
-[Unreleased]: https://github.com/Colin-b/pytest_httpx/compare/v0.11.0...HEAD
+[Unreleased]: https://github.com/Colin-b/pytest_httpx/compare/v0.12.0...HEAD
+[0.12.0]: https://github.com/Colin-b/pytest_httpx/compare/v0.11.0...v0.12.0
 [0.11.0]: https://github.com/Colin-b/pytest_httpx/compare/v0.10.1...v0.11.0
 [0.10.1]: https://github.com/Colin-b/pytest_httpx/compare/v0.10.0...v0.10.1
 [0.10.0]: https://github.com/Colin-b/pytest_httpx/compare/v0.9.0...v0.10.0